Renamed test, added Parsons label, and changed assert to include delta value
